work on HomOriented
